dom/tests/mochitest/cowl/postmessage/test_postmessage_structured_clone.html


<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <p id="display"></p>
  <pre id="test">

  <script class="testbody" type="application/javascript">
  SimpleTest.waitForExplicitFinish();

  var completedTests = 0;
  var passedTests = 0;

  function ok_wrapper() {
      ok(true, "Could pass privilege and label via postMessage");

      completedTests++;

      if (completedTests == 2) {
        // is(passedTests, completedTests, "There are " + completedTests + " privilege tests that should pass");

        SimpleTest.finish();
      }
  }

  function start() {
    var frame = document.getElementById('frame').contentWindow;
    // create
    var label = new Label('http://alice.com');
    frame.postMessage(label, '*');

    var priv = new FreshPrivilege();
    frame.postMessage(priv, '*');
  }

  </script>
  <iframe id="frame" onload="start()" src="file_postmessage_structured_clone.html"></iframe>
</pre>
</body>
</html>
